requirements.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
fake-useragent==1.1.1
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
oauthlib==3.2.2
parse==1.19.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.14
w3lib==2.1.1
websockets==10.4
zipp==3.11.0
psycopg2==2.9.5
uWSGI>=2.0.19.1,<2.1